Updated the spec to remove minitest references
Removed minitest references from the spec.
Fedora 19 requires signo to run on minitest 4.7.0
and doesnot have a suitable minitest rails package yet.
The tests are not getting run during the build stage anyway
so no point in including the minitest package. Also
every pull request in signo is tested via travis so
runnin tests again in the rpm is redundant.